+++ Editor/GraphicsSettingsManager.cs
@@ -62,11 +62,11 @@ namespace Kokowolo.Utilities.Editor
             // If it is null, the default is the Built-in Render Pipeline
             if (GraphicsSettings.defaultRenderPipeline != null)
             {
-                Debug.Log("The default render pipeline is defined by " + GraphicsSettings.defaultRenderPipeline.name);
+                LogManager.Log("The default render pipeline is defined by " + GraphicsSettings.defaultRenderPipeline.name);
             }
             else
             {
-                Debug.Log("The default render pipeline is the Built-in Render Pipeline");
+                LogManager.Log("The default render pipeline is the Built-in Render Pipeline");
             }
 
             // QualitySettings.renderPipeline determines the override render pipeline for the current quality level
@@ -74,24 +74,24 @@ namespace Kokowolo.Utilities.Editor
             // If an override render pipeline is defined, Unity uses that; otherwise, it falls back to the default value
             if (QualitySettings.renderPipeline != null)
             {
-                Debug.Log("The override render pipeline for the current quality level is defined by " + QualitySettings.renderPipeline.name);
-                Debug.Log("The active render pipeline is the override render pipeline");
+                LogManager.Log("The override render pipeline for the current quality level is defined by " + QualitySettings.renderPipeline.name);
+                LogManager.Log("The active render pipeline is the override render pipeline");
             }
             else
             {
-                Debug.Log("No override render pipeline exists for the current quality level");
-                Debug.Log("The active render pipeline is the default render pipeline");
+                LogManager.Log("No override render pipeline exists for the current quality level");
+                LogManager.Log("The active render pipeline is the default render pipeline");
             }
 
             // To get a reference to the Render Pipeline Asset that defines the active render pipeline,
             // without knowing if it is the default or an override, use GraphicsSettings.currentRenderPipeline
             if (GraphicsSettings.currentRenderPipeline != null)
             {
-                Debug.Log("The active render pipeline is defined by " + GraphicsSettings.currentRenderPipeline.name);
+                LogManager.Log("The active render pipeline is defined by " + GraphicsSettings.currentRenderPipeline.name);
             }
             else
             {
-                Debug.Log("The active render pipeline is the Built-in Render Pipeline");
+                LogManager.Log("The active render pipeline is the Built-in Render Pipeline");
             }
         }